+++ ostd/src/mm/kspace/mod.rs
@@ -188,12 +188,12 @@ unsafe impl PageTableConfig for KernelPtConfig {
         assert(nr_pte_index_bits::<PagingConsts>() == 9_usize);
         assert(PagingConsts::BASE_PAGE_SIZE().ilog2() == 12u32);
         assert(pte_index_bit_offset_spec::<PagingConsts>(4) == 39);
-        assert((256 as int) * pow2(39) == pow2(47));
-        assert(((256 as int) * (pow2(39) as int)) / (pow2(47) as int) == 1);
+        assert(256 * pow2(39) == pow2(47));
+        assert((256 * pow2(39) as int) / (pow2(47) as int) == 1);
         assert(pte_index_bit_offset_spec::<Self::C>(Self::C::NR_LEVELS()) == 39);
-        assert((Self::TOP_LEVEL_INDEX_RANGE_spec().start as int) * (pow2(
+        assert((Self::TOP_LEVEL_INDEX_RANGE_spec().start) * (pow2(
             pte_index_bit_offset_spec::<Self::C>(Self::C::NR_LEVELS()) as nat,
-        ) as int) == pow2(47) as int);
+        )) == pow2(47));
         assert((((Self::TOP_LEVEL_INDEX_RANGE_spec().start as int) * (pow2(
             pte_index_bit_offset_spec::<Self::C>(Self::C::NR_LEVELS()) as nat,
         ) as int)) / (pow2((Self::C::ADDRESS_WIDTH() - 1) as nat) as int)) == 1);
@@ -201,8 +201,8 @@ unsafe impl PageTableConfig for KernelPtConfig {
             pte_index_bit_offset_spec::<Self::C>(Self::C::NR_LEVELS()) as nat,
         ) as int)) / (pow2((Self::C::ADDRESS_WIDTH() - 1) as nat) as int)) % 2 == 1);
         lemma_pow2_adds(16, 48);
-        assert(Self::LEADING_BITS_spec() as int * 0x1_0000_0000_0000int
-            == 0x1_0000_0000_0000_0000int - pow2(Self::C::ADDRESS_WIDTH() as nat) as int);
+        assert(Self::LEADING_BITS_spec() * 0x1_0000_0000_0000int == 0x1_0000_0000_0000_0000int
+            - pow2(Self::C::ADDRESS_WIDTH() as nat));
     }
 
     fn TOP_LEVEL_INDEX_RANGE() -> (r: Range<usize>) {
