+++ ostd/src/mm/page_table/cursor/mod.rs
@@ -1825,7 +1825,7 @@ impl<'rcu, C: PageTableConfig, A: InAtomicMode> Cursor<'rcu, C, A> {
         let taken = self.path[self.level as usize - 1].take().unwrap_or_panic();
         proof {
             let ghost child_cont = owner.continuations[owner.level - 1];
-            assert(old(self).path[old(self).level as int - 1] is Some);
+            assert(old(self).path[old(self).level - 1] is Some);
             assert(child_cont.all_some());
             assert(child_cont.inv());
             assert(taken == owner.continuations[owner.level - 1].guard);
@@ -1944,7 +1944,7 @@ impl<'rcu, C: PageTableConfig, A: InAtomicMode> Cursor<'rcu, C, A> {
             old(owner).in_locked_range(),
             old(self).wf(*old(owner)),
             1 <= old(self).level <= 4,
-            old(self).path[old(self).level as int - 1] is Some,
+            old(self).path[old(self).level - 1] is Some,
             old(owner).metaregion_sound(*regions),
             regions.inv(),
         ensures
@@ -1955,19 +1955,19 @@ impl<'rcu, C: PageTableConfig, A: InAtomicMode> Cursor<'rcu, C, A> {
             final(self).guard_level == old(self).guard_level,
             final(self).va == old(self).va,
             final(self).barrier_va == old(self).barrier_va,
-            *res.node == old(self).path[old(self).level as int - 1]->0,
-            final(self).path[old(self).level as int - 1] is Some,
-            final(self).path[old(self).level as int - 1]->0 == *final(res.node),
-            *final(res.node) == old(self).path[old(self).level as int - 1]->0
-                ==> final(self).path[old(self).level as int - 1]->0
-                    == old(self).path[old(self).level as int - 1]->0,
-            *final(res.node) == *res.node ==> final(self).path[old(self).level as int - 1]->0
+            *res.node == old(self).path[old(self).level - 1]->0,
+            final(self).path[old(self).level - 1] is Some,
+            final(self).path[old(self).level - 1]->0 == *final(res.node),
+            *final(res.node) == old(self).path[old(self).level - 1]->0
+                ==> final(self).path[old(self).level - 1]->0
+                    == old(self).path[old(self).level - 1]->0,
+            *final(res.node) == *res.node ==> final(self).path[old(self).level - 1]->0
                 == *res.node,
-            *final(res.node) == old(self).path[old(self).level as int - 1]->0
+            *final(res.node) == old(self).path[old(self).level - 1]->0
                 ==> final(self).wf(*final(owner)),
             *final(res.node) == *res.node ==> final(self).wf(*final(owner)),
             forall|i: int|
-                0 <= i < NR_LEVELS && i != old(self).level as int - 1 ==> #[trigger] final(self).path[i]
+                0 <= i < NR_LEVELS && i != old(self).level - 1 ==> #[trigger] final(self).path[i]
                     == old(self).path[i],
             *final(owner) == *old(owner),
             final(owner).metaregion_sound(*regions),
@@ -2117,10 +2117,10 @@ impl<'rcu, C: PageTableConfig, A: InAtomicMode> CursorMut<'rcu, C, A> {
     fn protect_cur_entry(&mut self, op: impl FnOnce(PageProperty) -> PageProperty) {
         proof {
             assert(self.0.wf(*owner));
-            assert(self.0.path[self.0.level as int - 1] is Some);
+            assert(self.0.path[self.0.level - 1] is Some);
         }
         let entry_idx = pte_index::<C>(self.0.va, self.0.level);
-        let ghost cur_path_guard = self.0.path[self.0.level as int - 1]->0;
+        let ghost cur_path_guard = self.0.path[self.0.level - 1]->0;
 
         let ghost owner0 = *owner;
 
@@ -2155,7 +2155,7 @@ impl<'rcu, C: PageTableConfig, A: InAtomicMode> CursorMut<'rcu, C, A> {
             assert(owner.continuations[owner.level - 1].all_some()) by {
                 let cont_new = owner.continuations[owner.level - 1];
                 assert forall|i: int| 0 <= i < NR_ENTRIES implies cont_new.children[i] is Some by {
-                    if i != cont_new.idx as int {
+                    if i != cont_new.idx {
                         assert(cont0.children[i] is Some);
                     }
                 };
@@ -2422,7 +2422,7 @@ impl<'rcu, C: PageTableConfig, A: InAtomicMode> CursorMut<'rcu, C, A> {
     fn map_branch_pt(&mut self, pt: PageTableNodeRef<'rcu, C>, rcu_guard: &'rcu A) {
         let ghost guards0 = *guards;
 
-        let ghost level_key = owner.level as int - 1;
+        let ghost level_key = owner.level - 1;
         let ghost cont_orig = owner.continuations[level_key];
 
         let tracked mut continuation = owner.continuations.tracked_remove(owner.level - 1);
@@ -2627,10 +2627,10 @@ impl<'rcu, C: PageTableConfig, A: InAtomicMode> CursorMut<'rcu, C, A> {
                     let ghost guards_pre_none = *guards;
                     proof {
                         assert(self.0.wf(owner1));
-                        assert(self.0.path[cur_level as int - 1] is Some);
+                        assert(self.0.path[cur_level - 1] is Some);
                     }
                     let entry_idx = pte_index::<C>(self.0.va, self.0.level);
-                    let ghost cur_path_guard = self.0.path[cur_level as int - 1]->0;
+                    let ghost cur_path_guard = self.0.path[cur_level - 1]->0;
 
                     let tracked mut continuation = owner.continuations.tracked_remove(
                         owner.level - 1,
@@ -2642,10 +2642,9 @@ impl<'rcu, C: PageTableConfig, A: InAtomicMode> CursorMut<'rcu, C, A> {
 
                     proof {
                         assert(owner_pre_none == owner1);
-                        assert(cont_pre_alloc == owner1.continuations[cur_level as int - 1]);
-                        assert(entry_idx == AbstractVaddr::from_vaddr(
-                            self.0.va,
-                        ).index[cur_level as int - 1]);
+                        assert(cont_pre_alloc == owner1.continuations[cur_level - 1]);
+                        assert(entry_idx == AbstractVaddr::from_vaddr(self.0.va).index[cur_level
+                            - 1]);
                         assert(entry_idx == cont_pre_alloc.idx);
                         assert(cur_path_guard == cont_pre_alloc.guard);
                         assert(parent_owner.relate_guard(cur_path_guard));
@@ -2781,7 +2780,7 @@ impl<'rcu, C: PageTableConfig, A: InAtomicMode> CursorMut<'rcu, C, A> {
                             assert(cont_new.children[cont_new.idx as int] is Some);
                             assert forall|i: int|
                                 0 <= i < NR_ENTRIES implies cont_new.children[i] is Some by {
-                                if i != cont_new.idx as int {
+                                if i != cont_new.idx {
                                     assert(cont_pre_alloc.children[i] is Some);
                                 }
                             };
@@ -3113,7 +3112,7 @@ impl<'rcu, C: PageTableConfig, A: InAtomicMode> CursorMut<'rcu, C, A> {
 
         proof {
             // Hoist common facts about new_owner from new_child postcondition.
-            let cont = owner1.continuations[owner1.level as int - 1];
+            let cont = owner1.continuations[owner1.level - 1];
             assert(new_owner.value == EntryOwner::<C>::new_frame(
                 pa,
                 cont.path().push_tail(cont.idx as usize),
@@ -3944,10 +3943,10 @@ impl<'rcu, C: PageTableConfig, A: InAtomicMode> CursorMut<'rcu, C, A> {
 
         proof {
             assert(self.0.wf(owner0));
-            assert(self.0.path[level as int - 1] is Some);
+            assert(self.0.path[level - 1] is Some);
         }
         let entry_idx = pte_index::<C>(self.0.va, self.0.level);
-        let ghost cur_path_guard = self.0.path[level as int - 1]->0;
+        let ghost cur_path_guard = self.0.path[level - 1]->0;
 
         let tracked mut continuation = owner.continuations.tracked_remove((owner.level - 1) as int);
         let ghost cont0 = continuation;
@@ -3969,8 +3968,8 @@ impl<'rcu, C: PageTableConfig, A: InAtomicMode> CursorMut<'rcu, C, A> {
         let ghost old_child_pre_replace = old_child_owner.value;
 
         proof {
-            assert(cont0 == owner0.continuations[(owner0.level - 1) as int]);
-            assert(entry_idx == AbstractVaddr::from_vaddr(self.0.va).index[level as int - 1]);
+            assert(cont0 == owner0.continuations[owner0.level - 1]);
+            assert(entry_idx == AbstractVaddr::from_vaddr(self.0.va).index[level - 1]);
             assert(entry_idx == cont0.idx);
             assert(cur_path_guard == cont0.guard);
             cont0.inv_children_rel_unroll(cont0.idx as int);
@@ -4167,7 +4166,7 @@ impl<'rcu, C: PageTableConfig, A: InAtomicMode> CursorMut<'rcu, C, A> {
                         0 <= j < NR_ENTRIES
                             && final_cont.children[j] is Some implies final_cont.children[j].unwrap().tree_predicate_map(
                     final_cont.path().push_tail(j as usize), g_sound) by {
-                        if j != idx as int {
+                        if j != idx {
                             cont0.inv_children_unroll(j);
                             let subtree = cont0.children[j].unwrap();
                             let path_j = final_cont.path().push_tail(j as usize);
